requirements.txt
bayesian-optimization==1.5.1
joblib==1.4.2
matplotlib==3.9.1
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.1
scikit-learn==1.5.0
ta==0.11.0
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
xgboost==2.1.0rc1
yfinance==0.2.40